backend/requirements.txt
Django==3.2.12
django-cors-headers==3.7.0
django-environ==0.11.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
opencv-python==4.9.0.80
Pillow==8.3.1
psycopg2-binary==2.9.9
tensorflow==2.13.1
django-dotenv
python-dotenv
boto3
cloudinary